@@ -1,10 +1,9 @@
-import React, { ReactNode, Component } from "react";
+import React from "react";
 import { ContactInput } from "./ContactInput";
 import { ContactSubmit } from "./ContactSubmit";
-import { IRedirectService } from "../../services/RedirectService";
+import { useNavigate } from "react-router-dom";
 
 interface Props {
-  redirector: IRedirectService;
   name?: string;
   email?: string;
   subject?: string;
@@ -28,59 +27,56 @@ interface sendmailResponse {
   invalidEmail?: string;
 }
 
-export class ContactForm extends Component<Props, State> {
-  readonly state: State = {
+function hasAllContent(state: State): boolean {
+  return !!state.name && !!state.email && !!state.subject && !!state.message;
+}
+
+function validateContent(content: string): boolean {
+  return !!content;
+}
+function validateEmail(email: string): boolean {
+  var re = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+  return re.test(String(email).toLowerCase());
+}
+
+export function ContactForm(props: Props) {
+  const navigate = useNavigate();
+  const [state, setState] = React.useState<State>({
     canSubmit: false,
     processing: false,
-    name: this.props.name || "",
-    email: this.props.email || "",
-    subject: this.props.subject || "",
-    message: this.props.message || ""
-  };
+    name: props.name || "",
+    email: props.email || "",
+    subject: props.subject || "",
+    message: props.message || ""
+  });
 
-  private hasAllContent(state: State): boolean {
-    return !!state.name && !!state.email && !!state.subject && !!state.message;
-  }
-
-  private validateContent(content: string): boolean {
-    return !!content;
-  }
-  private validateEmail(email: string): boolean {
-    var re = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-    return re.test(String(email).toLowerCase());
-  }
-
-  private onInputChange(
-    name: keyof State,
-    value: string,
-    inError: boolean
-  ): void {
-    let newState = Object.assign(this.state, {
+  const onInputChange = (name: keyof State, value: string, inError: boolean): void => {
+    let newState = Object.assign(state, {
       [name]: value
     }) as State;
     newState = {
       ...newState,
-      canSubmit: !inError && this.hasAllContent(newState)
+      canSubmit: !inError && hasAllContent(newState)
     } as Pick<State, keyof State>;
-    this.setState(newState);
+    setState(newState);
   }
 
-  private async onSubmit(
+  const onSubmit = async(
     event: React.MouseEvent<HTMLButtonElement, MouseEvent>
-  ): Promise<void> {
+  ): Promise<void> => {
     event.preventDefault();
-    if (!this.state.canSubmit) {
+    if (!state.canSubmit) {
       return;
     }
 
     const response = await fetch("sendmail.php", {
       method: "POST",
       body: (
-        `name=${encodeURIComponent(this.state.name)}&` +
-        `email=${encodeURIComponent(this.state.email)}&` +
-        `subject=${encodeURIComponent(this.state.subject)}&` +
-        `message=${encodeURIComponent(this.state.message)}&` +
-        `context=${encodeURIComponent(this.props.context || "")}`
+        `name=${encodeURIComponent(state.name)}&` +
+        `email=${encodeURIComponent(state.email)}&` +
+        `subject=${encodeURIComponent(state.subject)}&` +
+        `message=${encodeURIComponent(state.message)}&` +
+        `context=${encodeURIComponent(props.context || "")}`
       ).replace(/%20/g, "+"),
       headers: {
         "Content-Type": "application/x-www-form-urlencoded"
@@ -88,67 +84,65 @@ export class ContactForm extends Component<Props, State> {
     });
 
     if (response.ok) {
-      this.props.redirector.goTo("/thankyou");
+      navigate("/thankyou");
     } else {
       try {
         const data: sendmailResponse = await response.json();
-        this.props.redirector.goTo("/error", { message: data.message });
+        navigate("/error", { state: { message: data.message} });
       } catch (err) {
         console.log(err);
-        this.props.redirector.goTo(`/error`);
+        navigate(`/error`);
       }
     }
   }
 
-  public render(): ReactNode {
-    return (
-      <form action="sendmail.php" method="Post">
-        <ContactInput
-          name="name"
-          label="Full Name: "
-          value={this.state.name}
-          error="Please supply your name"
-          validation={this.validateContent.bind(this)}
-          onChange={(err, val) =>
-            this.onInputChange.bind(this)("name", val, err)
-          }
-        />
-        <ContactInput
-          name="email"
-          label="Email Address: "
-          value={this.state.email}
-          error="Please supply a valid email address"
-          validation={this.validateEmail.bind(this)}
-          onChange={(err, val) =>
-            this.onInputChange.bind(this)("email", val, err)
-          }
-        />
-        <ContactInput
-          name="subject"
-          label="Subject: "
-          value={this.state.subject}
-          error="Please supply a subject"
-          validation={this.validateContent.bind(this)}
-          onChange={(err, val) =>
-            this.onInputChange.bind(this)("subject", val, err)
-          }
-        />
-        <ContactInput
-          name="message"
-          label="Message: "
-          value={this.state.message}
-          error="Please type a message"
-          validation={this.validateContent.bind(this)}
-          multiline={true}
-          onChange={(err, val) =>
-            this.onInputChange.bind(this)("message", val, err)
-          }
-        />
-        <ContactSubmit
-          onClick={this.onSubmit.bind(this)}
-          canSubmit={this.state.canSubmit}
-        ></ContactSubmit>
-      </form>
-    );
-  }
+  return (
+    <form action="sendmail.php" method="Post">
+    <ContactInput
+      name="name"
+      label="Full Name: "
+      value={state.name}
+      error="Please supply your name"
+      validation={validateContent}
+      onChange={(err, val) =>
+        onInputChange("name", val, err)
+      }
+    />
+    <ContactInput
+      name="email"
+      label="Email Address: "
+      value={state.email}
+      error="Please supply a valid email address"
+      validation={validateEmail}
+      onChange={(err, val) =>
+        onInputChange("email", val, err)
+      }
+    />
+    <ContactInput
+      name="subject"
+      label="Subject: "
+      value={state.subject}
+      error="Please supply a subject"
+      validation={validateContent}
+      onChange={(err, val) =>
+        onInputChange("subject", val, err)
+      }
+    />
+    <ContactInput
+      name="message"
+      label="Message: "
+      value={state.message}
+      error="Please type a message"
+      validation={validateContent}
+      multiline={true}
+      onChange={(err, val) =>
+        onInputChange("message", val, err)
+      }
+    />
+    <ContactSubmit
+      onClick={onSubmit}
+      canSubmit={state.canSubmit}
+    ></ContactSubmit>
+  </form>
+);
 }
