@@ -1,9 +1,9 @@
-import React, { ReactNode, Component } from "react";
+import React, { ReactNode, Component, PropsWithChildren } from "react";
 import "./MarginCue.css";
 
 interface Props {}
 
-export class Name extends Component<Props> {
+export class Name extends Component<PropsWithChildren<Props>> {
   public render(): ReactNode {
     return <span className="named">{this.props.children}</span>;
   }
